controle de acesso e validacoes de usuario e post

+++ app/Http/Controllers/PostsController.php
@@ -11,6 +11,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -92,6 +102,11 @@ class PostsController extends Controller
     {
         $post = Post::find($id);
 
+        // Check for correct user
+        if (auth()->user()-> !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -125,6 +140,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()-> !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         $post->delete();
 
         // Redirect and Set success message
